synthetic resolution using quote system (#252)
* Implement fragment provider and synthetic resolve extension with descriptors created on the transformations of the Quote system. The quote system is executed on the files of the current project when the editor changed or .kt files on disk changed.

This commit re-enables the synthetic resolve extension, which previously was disabled. The synthetic extension was extracted from the MetaSyntheticFragmentProvider. Previously, the fragment provider was implementing two extension points and therefore would have been instantiated twice. Now fragment provider and synthetic resolve extension are distinct implementations. They both access the cache maintained by QuoteSystemCache.

This adds an early test to test that resolving works for static content and when a user types into the editor.

* Remove unneeded changes to imports, remove unused class

* Apply new quote transformation to match implementation of master

* faster retrieval of .kt files of the current project

* Only provide package fragment descriptors for the queried package, not all packages available in the cache

* Small optimizations, added fixme notes

* chore: cleanup documentation

* improvements to cache update:
- perform initial transformation of all project .kt files in a background task. This tells the user what's happening
- don't transform all project files after a file change. Index access is prohibited in 2019.3 and incremental updates are better, anyway
- a few smaller improvements to the cache handling

* Improve the simple implementation for an incrementally update the cache of transformation results
Fix initial highlighting after startup. It was showing errors before.

* Refactor QuoteSystemCache:
extract the cache update handling into a separate interface and class. In the future we'll run into more issues with incremental cache updates and being able to test this part easily will help to fix this.
Also, the caching will change, too. For example, to not keep strong references to PsiFiles. Having this in a strictly defined place helps to maintain this.

This also adds more test coverage for incremental cache updates and fixes a bug, which occurred when nothing was transformed after an update for a file.

* Cleanup and update documentation

* Fix tests. A transformation started in the DocumentChangeListener was using the uncommitted PsiFile. This meant, that the text applied by the editor modification isn't available yet in the PsiFile, but only in the editor and the document instance.

Co-authored-by: Raúl Raja Martínez <raulraja@gmail.com>